[modify] change the structure
